Update javascript to take user min/max input
The minimum field and maximum field next to the guess field now take user input. On a keyup event the minimum or maximum is updated and the new min/max is displayed dynamically. Also every keyup generates a new random number, to ensure it accomodates the new range. The range also updates every time a user enters a new correct guess, which is also displayed.