Tweaks to be phase 3 ready
Added and cleaned up some javascript functionality, including a classList.add event that highlights when the range changes on the first correct guess. A small thing that highlights the subtle change that winning the guess has. Added some ARIA functionality to HTML, but I know there is more to be done there. Also changed the minimum and maximum input fields to update variables based on the 'input' event listener and not 'keyup,' which helps people tab through the app. Before that change it was counting 'tab' key presses as the new variable value, which was messing up all sorts of things.